[GH-2] Move window procedure to separate module in order to make the code cleaner.
